// visus-front/packages/api/src/models/UpdateGestorDTO.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const updateGestorDTOGeneroEnum = {
    "MASCULINO": "MASCULINO",
    "FEMININO": "FEMININO",
    "OUTRO": "OUTRO",
    "NAO_INFORMAR": "NAO_INFORMAR"
} as const;

export type UpdateGestorDTOGeneroEnumKey = (typeof updateGestorDTOGeneroEnum)[keyof typeof updateGestorDTOGeneroEnum];

export type UpdateGestorDTO = {
    /**
     * @type string | undefined
    */
    nome?: string;
    /**
     * @type string | undefined
    */
    telefone?: string;
    /**
     * @type string | undefined
    */
    genero?: UpdateGestorDTOGeneroEnumKey;
    /**
     * @type string | undefined, date
    */
    dataNascimento?: string;
    /**
     * @type string | undefined
    */
    matricula?: string;
    /**
     * @type string | undefined
    */
    cargo?: string;
};